Move all starters to same package
